requirements.txt
torch>=1.10
torchvision>=0.11
numpy>=1.21
Pillow>=8.4
tqdm>=4.62
matplotlib>=3.4
scikit-image>=0.18
tensorboard>=2.7
opencv-python>=4.5
